Fix file names, fix test and fix py formatting
